requirements.txt
azure-storage-blob>=12.19.0
fastapi>=0.110.0
google-generativeai>=0.3.1
Pillow>=10.1.0
PyPDF2>=3.0.1
pdf2image>=1.16.3
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
SQLAlchemy>=2.0.23
uvicorn[standard]>=0.27.0
pymongo>=4.6.0
motor>=3.3.2
